refactor(routes): :art: Added post route handler
Using `/:id` on posts could match unintended routes
