+++ server.js
@@ -9,21 +9,14 @@ const { Server } = require('socket.io')
 const { AddFollowing } = require('./controllers/AddFollowing')
 const { GetFollowing } = require('./controllers/GetFollowing')
 const { GetFollowers } = require('./controllers/GetFollowers')
-const { DeleteFollowing } = require('./controllers/DeleteFollowing')
-const { AddPost } = require('./controllers/AddPost')
-const { GetPosts } = require('./controllers/GetPosts');
+const { DeleteFollowing } = require("./controllers/DeleteFollowing");
 const { UserRouter } = require("./routes/UserRoute");
 const { FeedRouter } = require('./routes/Feed');
+const { Auth } = require("./routes/Auth");
+const { GetUsers } = require("./routes/UserRoute/GetUsers");
+const { PostRouter } = require("./routes/PostRoute");
 
-const AddLikes = require('./controllers/AddLikes').AddLikes;
-const DeleteLikes = require('./controllers/DeleteLikes').DeleteLikes;
-
-
-const CheckFollowing = require('./controllers/CheckFollowing').CheckFollowing
-
-const Auth = require("./routes/Auth").Auth;
-const GetUsers = require("./routes/UserRoute/GetUsers").GetUsers;
-const AddRepost = require('./controllers/AddRepost').AddRepost;
+const CheckFollowing = require("./controllers/CheckFollowing").CheckFollowing;
 
 const PORT = process.env.PORT || 4000
 const app = express();
@@ -38,23 +31,21 @@ server.listen(PORT);
 
 app.use("/auth", Auth);
 app.post('/addfollowing', (req, res) => { AddFollowing(req, res) });
+// RFC: The next two routes are the same. Should they be merged?
+// Or rather, should the entire following/followers be merged into the user object? 
+// since it'll mostly be accessed in that context. Hence making this part of the user route.
 app.get('/getfollowing/:user_id', (req, res) => { GetFollowing(req, res) });
-app.get('/getfollowers', (req, res)=> { GetFollowers(req, res) });
+app.get('/getfollowers/:user_id', (req, res) => {GetFollowers(req, res);});
 app.delete('/deletefollowing', (req, res) => { DeleteFollowing(req,res) });
 app.get('/checkfollowing/:user_id/:following', (req, res)=> CheckFollowing(req, res));
 
-app.post('/addpost', (req, res)=> { AddPost(req, res) });
-app.get('/getposts/:id', (req, res)=> { GetPosts(req, res)})
 
 app.use("/user", UserRouter);
 app.get('/getusers/:text', (req, res)=> { GetUsers(req, res) })
 
-app.patch('/addlikes', (req, res)=> { AddLikes(req, res) })
-app.delete('/deletelikes', (req, res)=> { DeleteLikes(req, res) })
 
-app.post('/addrepost', (req, res)=> { AddRepost(req, res) })
-
-app.use('/feed',FeedRouter);
+app.use("/feed", FeedRouter);
+app.use("/post", PostRouter);
 
 
 //Realtime Chat Service
